ADD: Create forms folder and move forms.py
